DERBY-4169: Change two SMALLINT routine args to be INT so that they will match the corresponding int args of the target Java method.
git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@766722 13f79535-47bb-0310-9956-ffa450edef68

+++ java/testing/org/apache/derbyTesting/system/oe/schema/dataroutines.sql
@@ -19,7 +19,7 @@
 -- Function to provided an updated C_DATA column for a customer account.
 -- Section 2.5.2.2
 CREATE FUNCTION BAD_CREDIT_DATA(
-    C_DATA VARCHAR(500), W SMALLINT, D SMALLINT,
+    C_DATA VARCHAR(500), W INT, D INT,
     CW SMALLINT, CD SMALLINT, C_ID INTEGER,
     AMOUNT DECIMAL(6, 2))
 RETURNS VARCHAR(500)
